Removed IRVCountyContestResult.java
