+++ server/eclipse-project/src/main/resources/us/freeandfair/corla/persistence/entity_classes
@@ -24,4 +24,3 @@ us.freeandfair.corla.model.Assertion
 us.freeandfair.corla.model.NEBAssertion
 us.freeandfair.corla.model.NENAssertion
 us.freeandfair.corla.model.IRVComparisonAudit
-us.freeandfair.corla.model.IRVCountyContestResult
